+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - The default port for the library has been changed from 15432 to 25432 to avoid clashing with instances of the CLI app.
 - `PostgresBuilder` no longer derives `Default` to avoid accidentally building nonsensical builders with no root pw.
+- Added `db_fixture` function for easier database creation and sharing.
 
 ## [0.3.0]